+++ src/main/java/sen/school/primaire/annee/AnneeRepository.java
@@ -2,7 +2,7 @@ package sen.school.primaire.annee;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnneeRepository extends JpaRepository<Annee, Integer> {
+public interface AnneeRepository extends JpaRepository<Annee, Long> {
 
 
 }
